Feat: Gathering life skills (mining, foraging) (#207)
* parser and packet handlers to deal with extractors

* separate mastery level for minining and foraging

* implemented rarity level

* life skill UI updates

* server check for mastery exp before harvesting

* implemented fatigue countr

* reverting shop hash change

* decrement counter on client side

* determined threshold for success rates

* use xml val to determine success rate bracket

* removed ordering

* parse masterydifferentialfactor for drop rate

* minor edit

+++ MapleServer2/Types/Levels.cs
@@ -120,7 +120,7 @@ namespace MapleServer2.Types
 
         public void GainMasteryExp(MasteryType type, long amount)
         {
-            MasteryExp masteryExp = MasteryExp.FirstOrDefault(x => x.Type == (byte) type);
+            MasteryExp masteryExp = MasteryExp.FirstOrDefault(x => x.Type == type);
 
             if (masteryExp == null) // add mastery to list
             {
